ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pdf2image==1.16.3
pillow==10.1.0
# Add requirements for giza-pidparser here later
# torch
# torchvision
